Restructure codebase into types/, io/, scoring/, ms2pip/ modules
Group files by responsibility: types/ for Python-facing data structures,
io/ for file parsing and format dispatch, scoring/ for feature computation,
ms2pip/ for future ms2pip-specific functionality. No logic changes.
