+++ src/ms2pip/mod.rs
@@ -1,0 +1,1 @@
+// Future: XGBoost feature vector computation for ms2pip
