+++ src/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod ms2_spectrum;
+pub mod precursor;
